requirements.txt
python-telegram-bot==20.7
gspread==6.0.0
oauth2client==4.1.3
requests==2.31.0
python-dotenv==1.0.1